inference/app/requirements.txt
torch
transformers
datasets
faiss-cpu
faiss-gpu
gradio
Pillow
requests
numpy
